synthetic-data/requirements.txt
azure-eventhub==5.15.1
pandas==2.3.3
faker==39.0.0